+++ apps/web/client/src/server/api/routers/domain/index.ts
@@ -1,0 +1,26 @@
+import { previewDomains, publishedDomains } from '@onlook/db';
+import { eq } from 'drizzle-orm';
+import { z } from 'zod';
+import { createTRPCRouter, protectedProcedure } from '../../trpc';
+import { previewRouter } from './preview';
+import { verificationRouter } from './verify';
+
+export const domainRouter = createTRPCRouter({
+    preview: previewRouter,
+    verification: verificationRouter,
+    getAll: protectedProcedure.input(z.object({
+        projectId: z.string(),
+    })).query(async ({ ctx, input }) => {
+        const preview = await ctx.db.query.previewDomains.findMany({
+            where: eq(previewDomains.projectId, input.projectId),
+        });
+        const published = await ctx.db.query.publishedDomains.findMany({
+            where: eq(publishedDomains.projectId, input.projectId),
+        });
+
+        return {
+            preview,
+            published,
+        }
+    }),
+});
