+++ packages/db/src/schema/project/domain/custom.ts
@@ -1,7 +1,7 @@
 import { relations } from 'drizzle-orm';
 import { boolean, pgTable, text, timestamp, uuid } from 'drizzle-orm/pg-core';
 import { publishedDomains } from './published';
-import { customDomainVerification } from './verfication';
+import { customDomainVerification } from './verification';
 
 export const customDomains = pgTable('custom_domains', {
     id: uuid('id').primaryKey().defaultRandom(),
