+++ packages/db/src/schema/project/domain/index.ts
@@ -1,4 +1,4 @@
 export * from './custom';
 export * from './preview';
 export * from './published';
-export * from './verfication';
+export * from './verification';
